# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Caller Tracking System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: agent
    description: Agent operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /agent/next-number:
    get:
      operationId: getNextNumber
      tags: [agent]
      summary: Get next uncalled phone number for agent
      responses:
        "200":
          description: Next phone number or null if none left
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NextNumberResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /agent/calls:
    post:
      operationId: submitCall
      tags: [agent]
      summary: Submit call feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallInput"
      responses:
        "201":
          description: Call submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getAgentCalls
      tags: [agent]
      summary: Get agent's own call logs
      responses:
        "200":
          description: Call logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallLog"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /agent/stats:
    get:
      operationId: getAgentStats
      tags: [agent]
      summary: Get agent's personal statistics
      responses:
        "200":
          description: Agent statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStats"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/agents:
    get:
      operationId: listAgents
      tags: [admin]
      summary: List all agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentUser"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAgent
      tags: [admin]
      summary: Create a new agent account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentUser"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/agents/{id}:
    patch:
      operationId: updateAgent
      tags: [admin]
      summary: Update agent info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: deactivateAgent
      tags: [admin]
      summary: Deactivate an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/agents/{id}/reset-password:
    post:
      operationId: resetAgentPassword
      tags: [admin]
      summary: Reset agent password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordReset"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/phone-numbers:
    get:
      operationId: listPhoneNumbers
      tags: [admin]
      summary: List all phone numbers
      responses:
        "200":
          description: Phone numbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PhoneNumber"
    post:
      operationId: addPhoneNumber
      tags: [admin]
      summary: Add single phone number to agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneNumberInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/phone-numbers/bulk:
    post:
      operationId: bulkAddPhoneNumbers
      tags: [admin]
      summary: Bulk add phone numbers to agent (newline-separated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneBulkInput"
      responses:
        "200":
          description: Bulk add result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResult"

  /admin/phone-numbers/{id}:
    delete:
      operationId: deletePhoneNumber
      tags: [admin]
      summary: Delete a phone number
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get all agents stats overview
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentStatRow"

  /admin/agent-detail/{id}:
    get:
      operationId: getAgentDetail
      tags: [admin]
      summary: Get detailed call logs for one agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetailResult"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    LoginResult:
      type: object
      required: [success, user]
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/CurrentUser"

    CurrentUser:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [agent, admin]

    NextNumberResult:
      type: object
      required: [done]
      properties:
        done:
          type: boolean
        phoneNumber:
          type: ["string", "null"]
        phoneNumberId:
          type: ["integer", "null"]
        remaining:
          type: ["integer", "null"]
        total:
          type: ["integer", "null"]

    CallInput:
      type: object
      required: [phoneNumberId, outcome]
      properties:
        phoneNumberId:
          type: integer
        phoneNumber:
          type: string
        outcome:
          type: string
          enum: [interested, will_buy, phone_off, no_answer, hung_up]
        notes:
          type: ["string", "null"]

    CallLog:
      type: object
      required: [id, outcome, phoneNumber, recordedAt]
      properties:
        id:
          type: integer
        phoneNumber:
          type: string
        outcome:
          type: string
        notes:
          type: ["string", "null"]
        recordedAt:
          type: string

    AgentStats:
      type: object
      required: [total, interested, will_buy, phone_off, no_answer, hung_up, remaining]
      properties:
        total:
          type: integer
        interested:
          type: integer
        will_buy:
          type: integer
        phone_off:
          type: integer
        no_answer:
          type: integer
        hung_up:
          type: integer
        remaining:
          type: integer
        totalPhones:
          type: integer

    AgentUser:
      type: object
      required: [id, username, role, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
        isActive:
          type: boolean
        totalCalls:
          type: integer
        createdAt:
          type: string

    AgentInput:
      type: object
      required: [username, password, name]
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]

    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        isActive:
          type: boolean

    PasswordReset:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string

    PhoneNumber:
      type: object
      required: [id, phoneNumber, agentId, isActive]
      properties:
        id:
          type: integer
        phoneNumber:
          type: string
        agentId:
          type: integer
        agentName:
          type: ["string", "null"]
        campaign:
          type: ["string", "null"]
        isActive:
          type: boolean
        calledCount:
          type: integer
        lastCalledAt:
          type: ["string", "null"]
        createdAt:
          type: string

    PhoneNumberInput:
      type: object
      required: [phoneNumber, agentId]
      properties:
        phoneNumber:
          type: string
        agentId:
          type: integer
        campaign:
          type: ["string", "null"]

    PhoneBulkInput:
      type: object
      required: [numbers, agentId]
      properties:
        numbers:
          type: string
          description: Newline-separated phone numbers
        agentId:
          type: integer
        campaign:
          type: ["string", "null"]

    BulkResult:
      type: object
      required: [added, skipped]
      properties:
        added:
          type: integer
        skipped:
          type: integer

    AgentStatRow:
      type: object
      required: [agentId, agentName, totalCalls, totalPhones]
      properties:
        agentId:
          type: integer
        agentName:
          type: string
        username:
          type: string
        totalCalls:
          type: integer
        totalPhones:
          type: integer
        interested:
          type: integer
        will_buy:
          type: integer
        phone_off:
          type: integer
        no_answer:
          type: integer
        hung_up:
          type: integer
        avgSecondsPerCall:
          type: ["number", "null"]
        lastCallAt:
          type: ["string", "null"]

    AgentDetailResult:
      type: object
      required: [agent, calls]
      properties:
        agent:
          $ref: "#/components/schemas/AgentUser"
        calls:
          type: array
          items:
            $ref: "#/components/schemas/CallLogDetail"

    CallLogDetail:
      type: object
      required: [id, outcome, phoneNumber, recordedAt]
      properties:
        id:
          type: integer
        phoneNumber:
          type: string
        outcome:
          type: string
        notes:
          type: ["string", "null"]
        recordedAt:
          type: string
        secondsSincePrev:
          type: ["integer", "null"]