# lib/api-zod/src/generated/types/agentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Caller Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentInput {
  username: string;
  password: string;
  name: string;
  /** @nullable */
  email?: string | null;
}