# lib/api-zod/src/generated/types/agentStatRow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Caller Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentStatRow {
  agentId: number;
  agentName: string;
  username?: string;
  totalCalls: number;
  totalPhones: number;
  interested?: number;
  will_buy?: number;
  phone_off?: number;
  no_answer?: number;
  hung_up?: number;
  /** @nullable */
  avgSecondsPerCall?: number | null;
  /** @nullable */
  lastCallAt?: string | null;
}