# lib/api-zod/src/generated/types/agentStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Caller Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentStats {
  total: number;
  interested: number;
  will_buy: number;
  phone_off: number;
  no_answer: number;
  hung_up: number;
  remaining: number;
  totalPhones?: number;
}