# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Caller Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentUpdate {
  name?: string;
  /** @nullable */
  email?: string | null;
  isActive?: boolean;
}