# lib/api-zod/src/generated/types/agentUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Caller Tracking System API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentUser {
  id: number;
  username: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  role: string;
  isActive: boolean;
  totalCalls?: number;
  createdAt?: string;
}